LayoutTests/inspector/elements/elements-panel-rewrite-href.html

<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
<html>
<head>
<link rel="stylesheet" src="/bogusSheet1.css">
<link rel="stylesheet" src="bogusSheet2.css">
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script>

function test()
{
    InspectorTest.expandElementsTree(step1);

    function step1()
    {
        var innerMapping = WebInspector.domAgent._idToDOMNode;

        for (var nodeId in innerMapping) {
            var node = innerMapping[nodeId];
            if (node.nodeName() === "LINK" || node.nodeName() === "SCRIPT") {
                var segments = [];
                var href = WebInspector.resourceURLForRelatedNode(node, node.getAttribute("src") || node.getAttribute("href"));
                if (!href) {
                    segments.push("<empty>");
                    continue;
                }
                var parsedURL = href.asParsedURL();
                if (!parsedURL)
                    testController.notifyDone("FAIL: no URL match for <" + href + ">");
                var split = parsedURL.path.split("/");
                for (var i = split.length - 1, j = 0; j < 3 && i >= 0; --i, ++j)
                    segments.push(split[i]);
                InspectorTest.addResult(segments.reverse());
            }
        }
        InspectorTest.completeTest();
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that src and href element targets are rewritten properly.
</p>

</body>
</html>
