LayoutTests/inspector/elements/elements-panel-rewrite-href-expected.txt

Tests that src and href element targets are rewritten properly.

null
,bogusSheet1.css
inspector,elements,bogusSheet2.css
tests,inspector,inspector-test.js
tests,inspector,elements-test.js
